Renamed folders so that fall 2017 is now the default, instead of spring 2017
