+++ hw2/logz.py
