+++ hw2/plot.py
