+++ hw4/cost_functions.py
