+++ hw4/dynamics.py
